+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Restore support to pass mocking information to the compiler
 - Disallow parameters mapped to a header to be passed via the body instead (#2151)
 - Handle skipped and unavailable as failures when calculating increments (#2184)
+- Constrain agent name to string values (#2172)
 
 # Release 2020.3 (2020-07-02)